moved get_info to backend.py
