+++ nx_parallel/utils/__init__.py
@@ -1,1 +1,2 @@
 from .chunk import *
+from .backend import *
